ендпойнт по задачам v2
